add: security to restaurant-service through a JWTFilter

+++ restaurant-service/src/main/java/com/gizasystems/restaurantservice/controllers/ItemController.java
@@ -6,9 +6,13 @@ import com.gizasystems.restaurantservice.service.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
+import static com.gizasystems.restaurantservice.util.helperFunctions.getOwnerId;
 
 @AllArgsConstructor
 @RestController
@@ -21,6 +25,8 @@ public class ItemController {
     // Build Add Item REST API
     @PostMapping
     public ResponseEntity<ItemDto> createItem(@RequestBody ItemDto itemDto){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
 
         ItemDto savedItem = itemService.createItem(itemDto);
         return new ResponseEntity<>(savedItem, HttpStatus.CREATED);
@@ -29,6 +35,9 @@ public class ItemController {
     // Build Get Item REST API
     @GetMapping("{id}")
     public ResponseEntity<ItemDto> getItemByID(@PathVariable("id") Long itemId){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         ItemDto itemDto = itemService.getItemById(itemId);
         return ResponseEntity.ok(itemDto);
     }
@@ -36,6 +45,9 @@ public class ItemController {
     // Build Get All Items REST API
     @GetMapping
     public ResponseEntity<List<ItemDto>> getAllItems(){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         var items = itemService.getAllItems();
         return ResponseEntity.ok(items);
     }
@@ -44,6 +56,9 @@ public class ItemController {
     @PutMapping("{id}")
     public ResponseEntity<ItemDto> updateItem(@PathVariable("id")Long itemId,
                                               @RequestBody ItemDto updatedItem){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         ItemDto itemDto = itemService.updateItem(itemId, updatedItem);
         return ResponseEntity.ok(itemDto);
     }
@@ -51,6 +66,9 @@ public class ItemController {
     // Build Delete Item REST API
     @DeleteMapping("{id}")
     public ResponseEntity<String> deleteItem(@PathVariable("id")Long itemId){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         itemService.deleteItem(itemId);
         return ResponseEntity.ok("Item deleted successfully!");
     }
