+++ restaurant-service/src/main/java/com/gizasystems/restaurantservice/controllers/RestaurantController.java
@@ -10,6 +10,8 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+import static com.gizasystems.restaurantservice.util.helperFunctions.getOwnerId;
+
 @AllArgsConstructor
 @RestController
 @RequestMapping("/restaurants")
@@ -20,6 +22,9 @@ public class RestaurantController {
     // Build Add Restaurant REST API
     @PostMapping
     public ResponseEntity<RestaurantDto> createRestaurant(@RequestBody RestaurantDto restaurantDto){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         RestaurantDto savedRestaurant = restaurantService.createRestaurant(restaurantDto);
         return new ResponseEntity<>(savedRestaurant, HttpStatus.CREATED);
     }
@@ -27,6 +32,9 @@ public class RestaurantController {
     // Build Get Restaurant REST API
     @GetMapping("{id}")
     public ResponseEntity<RestaurantDto> getRestaurantById(@PathVariable("id") Long restaurantId){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         RestaurantDto restaurantDto = restaurantService.getRestaurantById(restaurantId);
         return ResponseEntity.ok(restaurantDto);
     }
@@ -34,6 +42,9 @@ public class RestaurantController {
     // Build Get All Restaurants REST API
     @GetMapping
     public ResponseEntity<List<RestaurantDto>> getAllRestaurants(){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         var restaurants = restaurantService.getAllRestaurants();
         return ResponseEntity.ok(restaurants);
     }
@@ -42,6 +53,9 @@ public class RestaurantController {
     @PutMapping("{id}")
     public ResponseEntity<RestaurantDto> updateRestaurant(@PathVariable("id") Long restaurantId,
                                                           @RequestBody RestaurantDto updatedRestaurant){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         RestaurantDto restaurantDto = restaurantService.updateRestaurant(restaurantId, updatedRestaurant);
         return ResponseEntity.ok(restaurantDto);
     }
@@ -49,6 +63,9 @@ public class RestaurantController {
     // Build Delete Restaurant REST API
     @DeleteMapping("{id}")
     public ResponseEntity<String> deleteRestaurant(@PathVariable("id") Long restaurantId){
+        Long ownerId = getOwnerId();
+        if(ownerId == null) return ResponseEntity.badRequest().build();
+
         restaurantService.deleteRestaurant(restaurantId);
         return ResponseEntity.ok("Restaurant deleted successfully!");
     }
